pyproject.toml
[project]
name = "stock-analysis-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.14.2",
    "lxml>=6.0.2",
    "python-dotenv>=1.0.0",
    "torch>=2.9.1",
    "trafilatura>=2.0.0",
    "transformers>=4.57.1",
]
